// admcentro-web-main/src/interfaces/IAssurance.ts
// Generated by https://quicktype.io

export interface IAssuranceResponse {
    results: number;
    code: number;
    status: string;
    ok: boolean;
    data: IAssurance[];
}

export interface IAssurance {
    id: number;
    ContractId: number;
    fullName: string;
    address: string;
    phone: string;
    email: string;
    cuit: string;
    obs: string;
    createdAt: string;
    updatedAt: string;
    deletedAt: null | string;
}